requirements.txt
Adafruit-Blinka==8.0.2
adafruit-circuitpython-as7341==1.2.8
adafruit-circuitpython-busdevice==5.1.10
adafruit-circuitpython-register==1.9.10
adafruit-circuitpython-typing==1.7.1
adafruit-circuitpython-vl53l4cd==1.1.0
Adafruit-PlatformDetect==3.24.1
Adafruit-PureIO==1.1.9
appdirs==1.4.4
certifi==2022.6.15
charset-normalizer==2.0.12
circup==1.1.0
click==8.1.3
findimports==2.2.0
idna==3.3
pyftdi==0.54.0
pynmea2==1.18.0
pyserial==3.5
pyusb==1.2.1
requests==2.28.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
semver==2.13.0
smbus2==0.4.2
sparkfun-pi-servo-hat==0.9.0
sparkfun-qwiic==1.1.6
sparkfun-qwiic-adxl313==0.0.7
sparkfun-qwiic-alphanumeric==0.0.1
sparkfun-qwiic-as6212==0.0.2
sparkfun-qwiic-bme280==0.9.0
sparkfun-qwiic-button==2.0.1
sparkfun-qwiic-ccs811==0.9.4
sparkfun-qwiic-dual-encoder-reader==0.0.2
sparkfun-qwiic-eeprom==0.0.1
sparkfun-qwiic-gpio==0.0.2
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-icm20948==0.0.1
sparkfun-qwiic-joystick==0.9.0
sparkfun-qwiic-keypad==0.9.0
sparkfun-qwiic-kx13x==1.0.0
sparkfun-qwiic-led-stick==0.0.1
sparkfun-qwiic-max3010x==0.0.2
sparkfun-qwiic-micro-oled==0.10.0
sparkfun-qwiic-oled-base==0.0.2
sparkfun-qwiic-oled-display==0.0.2
sparkfun-qwiic-pca9685==0.9.1
sparkfun-qwiic-pir==0.0.4
sparkfun-qwiic-proximity==0.9.0
sparkfun-qwiic-relay==0.0.2
sparkfun-qwiic-rfid==2.0.0
sparkfun-qwiic-scmd==0.9.1
sparkfun-qwiic-serlcd==0.0.1
sparkfun-qwiic-sgp40==0.0.4
sparkfun-qwiic-soil-moisture-sensor==0.0.2
sparkfun-qwiic-tca9548a==0.9.0
sparkfun-qwiic-titan-gps==0.1.1
sparkfun-qwiic-twist==0.9.0
sparkfun-qwiic-vl53l1x==1.0.1
sparkfun-top-phat-button==0.0.2
sparkfun-ublox-gps==1.1.4
sysv-ipc==1.1.0
update-checker==0.18.0
urllib3==1.26.9
